refactor(tag-input): extract subcomponents for improved modularity
